pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.0",
    "python-fasthtml>=0.8.0",
    "requests>=2.32.3",
    "tqdm>=4.66.5",
    "uvicorn>=0.32.0",
]
